+++ dp_tools/config/schemas.py
@@ -31,6 +31,7 @@ runsheet = {
             "read1_path": pa.Column(str),
             "read2_path": pa.Column(str, required=False), # Expect if paired_end is True
             "Source Name": pa.Column(str),
+            "Has Tech Reps": pa.Column(bool),
         },
         # define checks at the DataFrameSchema-level
         checks=check_read2_path_populated_if_paired_end